Crate acid_io

A pared-down version of std::io usable in no_std contexts.

Re-exports

pub use byteorder;

Structs

BufReader

The BufReader<R> struct adds buffering to any reader.

BufWriter

Wraps a writer and buffers its output.

Bytes

An iterator over u8 values of a reader.

Cursor

A Cursor wraps an in-memory buffer and provides it with a Seek implementation.

Empty

Empty ignores any data written via Write, and will always be empty (returning zero bytes) when read via Read.

Error

The error type for I/O operations of the Read, Write, Seek, and associated traits.

IntoInnerError

An error returned by BufWriter::into_inner which combines an error that happened while writing out the buffer, and the buffered writer object which may be used to recover from the condition.

IoSlice

A buffer type used with Write::write_vectored.

IoSliceMut

A buffer type used with Read::read_vectored.

LineWriter

Wraps a writer and buffers output to it, flushing whenever a newline (0x0a, '\n') is detected.

Lines

An iterator over the lines of an instance of BufRead.

Repeat

A reader which yields one byte over and over and over and over and over and…

Sink

A writer which will move data into the void.

Split

An iterator over the contents of an instance of BufRead split on a particular byte.

Enums

ErrorKind

A list specifying general categories of I/O error.

SeekFrom

Enumeration of possible methods to seek within an I/O object.

Traits

BufRead

A BufRead is a type of Reader which has an internal buffer, allowing it to perform extra ways of reading.

Read

The Read trait allows for reading bytes from a source.

Seek

The Seek trait provides a cursor which can be moved within a stream of bytes.

Write

A trait for objects which are byte-oriented sinks.

Functions

copy

Copies the entire contents of a reader into a writer.

empty

Creates a value that is always at EOF for reads, and ignores all data written.

repeat

Creates an instance of a reader that infinitely repeats one byte.

sink

Creates an instance of a writer which will successfully consume all data.

Type Aliases

Result

A specialized Result type for I/O operations.